dev_requirements.txt
django-extensions==1.7.4
ipython==5.1.0
